Refactor package and method name
