+++ zanata-war/src/main/java/org/zanata/service/TranslationStateCache.java
@@ -40,12 +40,8 @@ public interface TranslationStateCache {
      * Informs the cache that a text flow has changed its state in a given
      * locale. (It's really a Text Flow Target state)
      *
-     * @param textFlowId
-     *            The id of the text flow that has changed state.
-     * @param localeId
-     *            The locale for which state has changed.
-     * @param newState
-     *            The new state after the change.
+     * @param event
+     *            The text flow target state event
      */
     void textFlowStateUpdated(TextFlowTargetStateEvent event);
 
@@ -60,7 +56,7 @@ public interface TranslationStateCache {
 
     /**
      * Return boolean of textFlowTarget has validation error against validation
-     * rules {@link org.zanata.webtrans.share.model.ValidationAction}
+     * rules {@link org.zanata.webtrans.shared.model.ValidationAction}
      *
      * @param textFlowTargetId
      * @param validationId
@@ -91,6 +87,4 @@ public interface TranslationStateCache {
      * @param localeId
      */
     void clearDocumentStatistics(Long documentId, LocaleId localeId);
-
-    void docStatsUpdated(DocStatsEvent event);
 }
