+++ zanata-war/src/main/java/org/zanata/webhook/events/DocumentMilestoneEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014, Red Hat, Inc. and individual contributors as indicated by the
+ * Copyright 2016, Red Hat, Inc. and individual contributors as indicated by the
  * @author tags. See the copyright.txt file in the distribution for a full
  * listing of individual contributors.
  *
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA, or see the FSF
  * site: http://www.fsf.org.
  */
-package org.zanata.events.webhook;
+package org.zanata.webhook.events;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -40,9 +40,12 @@ import org.zanata.events.WebhookEventType;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@JsonPropertyOrder({"project", "version", "docId", "locale", "editorDocumentUrl", "milestone"})
+@JsonPropertyOrder({"project", "version", "docId", "locale", "editorDocumentUrl", "milestone", "type"})
 @EqualsAndHashCode
 public class DocumentMilestoneEvent extends WebhookEventType {
+
+    private static final String type = "DocumentMilestoneEvent";
+
     /**
      * Target project slug.
      * {@link org.zanata.model.HProject#slug}
@@ -80,6 +83,6 @@ public class DocumentMilestoneEvent extends WebhookEventType {
 
     @Override
     public String getType() {
-        return this.getClass().getName();
+        return type;
     }
 }
